refactor: extract EPUB package presentation
